TP0 con tests

+++ lib/libdyn/src/shared.c
@@ -4,6 +4,6 @@
 
 int shared_lib_function(char* msg)
 {
-    printf("Hi! I'm a dynamic lib that receives %s, and return '%d'\n", msg, strlen(msg));
+    printf("Hi! I'm a dynamic lib that receives %s, and return '%d'\n", msg, (int)strlen(msg));
     return strlen(msg);
 }
